requirements.txt
selenium==4.21.0
pytest==8.2.0
allure-pytest==2.13.5
webdriver-manager==4.0.1